Implement partially

+++ src/ghcr.rs
@@ -1,14 +1,16 @@
 use std::collections::HashMap;
 use std::fs::File;
-use std::io::Read;
+use std::io::{BufRead, BufReader, Read};
 use std::path::Path;
 use std::process::Command;
 use std::{env, fs, str};
 
-use anyhow::{bail, Context, Result};
+use anyhow::{bail, Context as _, Result};
 use const_format::formatc;
+use data_encoding::HEXUPPER;
 use debug_print::debug_println as dprintln;
-use flate2::read::GzDecoder;
+use flate2::bufread::GzDecoder;
+use ring::digest::{Context, Digest, SHA256};
 use serde_json::json;
 use which::which;
 
@@ -79,12 +81,28 @@ impl Ghcr {
         Ok(image_name.to_string())
     }
 
+    fn sha256_digest<R: BufRead>(mut reader: GzDecoder<R>) -> Result<Digest> {
+        let mut context = Context::new(&SHA256);
+        let mut buffer = [0; 1024];
+
+        loop {
+            let count = reader.read(&mut buffer)?;
+            if count == 0 {
+                break;
+            }
+            context.update(&buffer[..count]);
+        }
+        Ok(context.finish())
+    }
+
     pub fn upload_oci_image(&self, target_file: &Path, name: &str, version: &str) -> Result<()> {
         let image_name = self.check_existence(name, version)?;
 
-        let dir_name = format!("{image_name}--{version}");
+        let dir_name = format!("{}--{version}", image_name.replace("/", "-"));
         let root = Path::new(&dir_name);
-        fs::remove_dir_all(root)?;
+        if root.exists() {
+            fs::remove_dir_all(root)?;
+        }
         fs::create_dir(root)?;
 
         let oci_image = oci::Image::new()?;
@@ -130,7 +148,7 @@ impl Ghcr {
             ),
         ]);
 
-        dprintln!("Uploading {target_file:?}");
+        dprintln!("Uploading {target_file:?} ...");
         let tar_gz_sha256 = oci::Image::write_tar_gz(target_file, blobs)?;
 
         let arch = "amd64"; // package must be built at least on x86_64
@@ -139,13 +157,13 @@ impl Ghcr {
         // get decompressed sha256 digest
         let tar_gz = File::open(target_file)?;
         let tar_gz_size = tar_gz.metadata()?.len();
-        let mut tar = GzDecoder::new(tar_gz);
-        let mut data = String::new();
-        tar.read_to_string(&mut data)?;
-        let tar_sha256 = sha256::digest(data);
+        let tar = GzDecoder::new(BufReader::new(tar_gz));
+        let tar_sha256 = Self::sha256_digest(tar)?;
 
+        dprintln!("Creating image config ...");
         let (config_json_sha256, config_json_size) =
-            oci_image.write_image_config(arch, os, &tar_sha256, blobs)?;
+            oci_image.write_image_config(arch, os, &HEXUPPER.encode(tar_sha256.as_ref()), blobs)?;
+        dprintln!("Done: creating image config");
 
         let descriptor_annotations = HashMap::<String, String>::from([(
             "org.opencontainers.image.ref.name".to_string(),
@@ -169,8 +187,10 @@ impl Ghcr {
             }],
             "annotations": package_annotations,
         });
+        dprintln!("Creating image manifest ...");
         let (manifest_json_sha256, manifest_json_size) =
             oci_image.write_image_manifest(&image_manifest, blobs)?;
+        dprintln!("Done: creating image manifest");
 
         let manifests = vec![json!({
             "mediaType": "application/vnd.oci.image.manifest.v1+json",
@@ -182,8 +202,12 @@ impl Ghcr {
             },
             "annotations": descriptor_annotations,
         })];
+        dprintln!("Creating image index ...");
         let (index_json_sha256, index_json_size) =
             oci_image.write_image_index(&manifests, &package_annotations, blobs)?;
+        dprintln!("Done: creating image index");
+
+        dprintln!("Creating index json ...");
         oci_image.write_index_json(
             &index_json_sha256,
             index_json_size,
@@ -193,6 +217,7 @@ impl Ghcr {
                 version.to_string(),
             )]),
         )?;
+        dprintln!("Done: creating index json");
 
         // TODO: --- upload_oci_image ---
